added all commands as an enum and read them in Lexer.c so now that i am exicuting i can use switch case to jump to the exicution did not remove the typing for other keywords so that i can type check them in execution

+++ src/Parser.c
@@ -211,7 +211,7 @@ struct parsedata *Parse(TokenNode **tokenlist){
                     new_state->statement=NULL;
                     new_state->back=(*parsTracer);
                     Key *new_key= malloc(sizeof(Key));
-                    int len =strlen((*listtracer)->token->string);
+                    int len =(int)strlen((*listtracer)->token->string);
                     new_key->size=len-1;
                     new_key->size=(new_key->size<32)?new_key->size: 32;
                     new_key->key= malloc(32);
